Agrega crear pedido y pagar pedido

+++ coffee-shop-api/domain/src/main/java/com/ceiba/coffee/model/dto/CoffeeDto.java
@@ -21,7 +21,7 @@ public class CoffeeDto {
                 coffee.getId(),
                 coffee.getName(),
                 coffee.getCategory().getId(),
-                coffee.getPrice().getAmount(),
+                (int)coffee.getPrice().getAmount(),
                 coffee.getUnits());
     }
